requirements.txt
python-dotenv
slack_bolt
psycopg2
tkinter
glob
tkinter.filedialog
